Cambio en métodos para el calculo de los elements

+++ database/migrations/2024_02_07_123606_add_fields_to_training_contracts_table.php
@@ -14,8 +14,8 @@ class AddFieldsToTrainingContractsTable extends Migration
     public function up()
     {
         Schema::table('training_contracts', function (Blueprint $table) {
-            $table->decimal('daily_hours_1')->nullable();
-            $table->decimal('daily_hours_2')->nullable();
+            $table->double('daily_hours_1')->nullable();
+            $table->double('daily_hours_2')->nullable();
         });
     }
